backend/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.13.1
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
asyncio==4.0.0
asyncpg==0.29.0
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
coverage==7.10.6
cryptography==45.0.7
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.104.1
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.7.0
greenlet==3.2.4
groq==0.32.0
grpcio==1.74.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.9
# hiredis==2.2.3
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.75
langchain-text-splitters==0.3.11
langsmith==0.4.21
libmagic==1.0
lxml==6.0.1
Mako==1.3.10
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.6.4
networkx==3.5
nltk==3.9.1
numpy==2.3.2
ollama==0.5.3
openai==1.102.0
openpyxl==3.1.5
orjson==3.11.3
packaging==25.0
pandas==2.3.2
passlib==1.7.4
pillow==11.3.0
pluggy==1.6.0
portalocker==3.2.0
propcache==0.3.2
protobuf==6.32.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.1.0
pydantic_core==2.33.2
pypdf==6.0.0
PyPDF2==3.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.0
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.6
python-slugify==8.0.1
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.15.1
redis==5.0.1
regex==2025.8.29
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==0.27.0
sympy==1.14.0
tenacity==9.1.2
text-unidecode==1.3
threadpoolctl==3.6.0
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
transformers==4.56.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.24.0
# uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
zstandard==0.24.0
